% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\docType{data}
\name{data_transform}
\alias{data_transform}
\title{Data Transformation and filtration}
\usage{
data(GTEXv7)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{transformation}{type of transformation method ''Log2' or 'Z-score' or 'quantile' or 'NA' (no) transformaiton}

\item{margin}{1 indicates rows, 2 indicates columns}
}
\description{
Transformation of data using one of the four mentioned methods and filtering based on median value
}
\examples{
transdf <- data_transform(df=GTEXv7[,3:10],transformation='log2', marg=2);
}
