% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTEX}
\alias{GTEX}
\title{Gene-Tissue Expression (GTEx) Dataset v7}
\format{
A data frame with 8000 rows and 54 columns:
}
\source{
\url{https://www.genome.gov/Funded-Programs-Projects/Genotype-Tissue-Expression-Project}
}
\usage{
GTEX
}
\description{
A dataset containing the expression of >42000 genes in 54 human tissues
}
\keyword{datasets}
