% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTEXv7}
\alias{GTEXv7}
\title{Gene-Tissue Expression (GTEx) Dataset v7}
\format{
A data frame with 42136 rows and 54 columns:
\describe{
\item{Description}{Gene names(feature identifiers)}
}
}
\source{
\url{https://www.genome.gov/Funded-Programs-Projects/Genotype-Tissue-Expression-Project}
}
\usage{
GTEXv7
}
\description{
A dataset containing the expression of >42000 genes in 54 human tissues
}
\keyword{datasets}
