% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logTCGA40}
\alias{logTCGA40}
\title{The Cancer Genome Atlas, Pan Cancer Study Dataset for selected genes(log transformed)}
\format{
A data frame with 40 rows and 10071 columns:
}
\usage{
logTCGA40
}
\description{
Log 2 transformed dataset containing the expression of
40 genes in 2000 human cancer samples(tissues or cells)
}
\keyword{datasets}
