% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\docType{data}
\name{pair_scatter}
\alias{pair_scatter}
\title{Pair Scatter}
\usage{
data(TCGA40)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{featurelist}{list of features in the sequence as it occurs in the Dataframe}

\item{feature1, feature2}{Features to be compared}
}
\description{
Scatter between two chosen features
}
\examples{
featgroup<-grepl( "RNASE",rownames(TCGA40)) #optional
pscatter <- pair_scatter(df=TCGA40,featurelist=rownames(TCGA40), feature1="RNASEH2A", feature2="RNASEH2C");
}
