% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\docType{data}
\name{pairwise_corr}
\alias{pairwise_corr}
\title{Pairwise Correlation}
\usage{
data(TCGA40)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{featurelist}{list of features in the sequence as it occurs in the Dataframe}

\item{featuregroup}{list of TRUE/FALSE values for features to be observed vs rest of the features}

\item{clusterorder}{Character, the ordering method of the correlation matrix.'original' for original order (default),'AOE' for the angular order of the eigenvectors, 'FPC' for the first principal component order, 'hclust' for the hierarchical clustering order, 'alphabet' for alphabetical order.}

\item{clustermethod}{Character, the ordering method of the correlation matrix.'original' for original order (default),'AOE' for the angular order of the eigenvectors, 'FPC' for the first principal component order, 'hclust' for the hierarchical clustering order, 'alphabet' for alphabetical order.}
}
\description{
Get Pair wise correlations for a given list of features
}
\examples{
pairwisedf<-t(TCGA40[,3:ncol(TCGA40)]) 
featgroup<-grepl( "RNASE",rownames(pairwisedf)) #optional
pairwisecorr <- pairwise_corr(df=pairwisedf,featurelist=rownames(pairwisedf), featuregroup=featgroup, clusterorder="hclust");
}
