% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA40}
\alias{TCGA40}
\title{The Cancer Genome Atlas, Pan Cancer Study Dataset for selected genes}
\format{
A data frame with 40 rows and 10071 columns:
}
\source{
\url{https://www.cbioportal.org/}
}
\usage{
TCGA40
}
\description{
A dataset containing the expression of
40 genes in 10071 human cancer samples(tissues or cells)
}
\keyword{datasets}
