% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{data_transform}
\alias{data_transform}
\title{Data Transformation and filtration
Transformation of data using one of the four mentioned methods and filtering based on median value}
\usage{
data_transform(df,transformation,featurelist, medianthres)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{transformation}{type of transformation method ''Log2' or 'Z-score' or 'quantile' or 'NA' (no) transformaiton}

\item{featurelist}{listing of feature names or ids to measure correlation between}

\item{medianthres}{(optional)features have median below this numerical value are filtered or removed}
}
\description{
Data Transformation and filtration
Transformation of data using one of the four mentioned methods and filtering based on median value
}
\examples{
transdf<- data_transform(df=GTEXv7[-1],transformation='log2', featurelist=GTEXv7$Description)
transdf<- data_transform(df=GTEXv7[-1],transformation='log2', featurelist=GTEXv7$Description,
                         medianthres=1)



}
