% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logTCGA40}
\alias{logTCGA40}
\title{Gene-Tissue Expression (GTEx) Dataset v7}
\format{
A data frame with 40 rows and 10071 columns:
\describe{
\item{rownames}{Gene names(feature identifiers)}
}
}
\usage{
logTCGA40
}
\description{
Log 2 transformed dataset containing the expression of
40 genes in 10071 human cancer samples(tissues or cells)
}
\keyword{datasets}
