% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{pair_scatter}
\alias{pair_scatter}
\title{Pair Scatter
Scatter between two chosen features.}
\usage{
pair_scatter(df,featurelist, feature1, feature2, corrmeth)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{featurelist}{list of features in the sequence as it occurs in the Dataframe}

\item{feature1, feature2}{Features to be compared}

\item{corrmeth}{Correlation method used. 'pearson' or 'spearman' (default='pearson')}
}
\description{
Pair Scatter
Scatter between two chosen features.
}
\examples{
data("TCGA40") #Load TCGAdata for 40 genes
data("logTCGA40") #Load log transformed TCGAdata for 40 genes
pscatter <- pair_scatter(df=TCGA40,featurelist=rownames(TCGA40),
                        feature1="RNASEH2A",feature2="PCNA", corrmeth='pearson')
pscatter <- pair_scatter(df=logTCGA40,featurelist=rownames(logTCGA40),
                        feature1="RNASEH2A", feature2="PCNA", corrmeth='pearson')

}
