% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{pair_scatter}
\alias{pair_scatter}
\title{Pair Scatter
Scatter between two chosen features.}
\usage{
pair_scatter(df,featurelist, feature1, feature2, corrmeth)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{featurelist}{list of features in the sequence as it occurs in the Dataframe}

\item{feature1, feature2}{Features to be compared}

\item{corrmeth}{Correlation method used. 'pearson' or 'spearman' (default='pearson')}

\item{col}{Color for scatter plot, default="#1f78b4"}
}
\description{
Pair Scatter
Scatter between two chosen features.
}
\examples{
pscatter <- pair_scatter(df=TCGA40,featurelist=rownames(TCGA40),
                         feature1="RNASEH2A",feature2="PCNA", corrmeth='pearson', col= "#1f78b4")
pscatter <- pair_scatter(df=logTCGA40,featurelist=rownames(logTCGA40),
                         feature1="RNASEH2A", feature2="PCNA", corrmeth='pearson' col= "#b45b1f")

}
