% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{pairwise_corr}
\alias{pairwise_corr}
\title{Pairwise Correlation}
\usage{
pairwise_corr(df,featurelist, featuregroup,visorder,corrmeth)
}
\arguments{
\item{df}{numerical dataframe with rows having series of values for a single feature}

\item{featurelist}{list of features in the sequence as it occurs in the Dataframe}

\item{featuregroup}{(optional)list of TRUE/FALSE values for features to be observed vs rest of the features(ordering/clustering is disabled)}

\item{visorder}{The ordering method of the correlation matrix.'original' for original order ,'AOE' for the angular order of the eigenvectors, 'FPC' for the first principal component order, 'hclust'(default) for the hierarchical clustering order, 'alphabet' for alphabetical order.}

\item{corrmeth}{Correlation method used. 'pearson' or 'spearman' (default='pearson')}
}
\description{
Get Pair wise correlations for a given list of features
}
\examples{
corr_pair<-pairwise_corr(df=logTCGA40,featurelist=rownames(logTCGA40), visorder="hclust")
featgroup<-grepl( "RNASE",rownames(logTCGA40)) #optional, a set of TRUE/FALSE of length featurelist
corr_pair<-pairwisecorr <- pairwise_corr(df=logTCGA40,featurelist=rownames(logTCGA40),
                                         featuregroup=featgroup)

}
