% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\docType{data}
\name{primefeature_corr}
\alias{primefeature_corr}
\title{Prime Feature Correlation}
\usage{
data(GTEXv7)
}
\arguments{
\item{df}{A dataframe with echo row corresponding to a feature}

\item{featurelist}{listing of feature names or ids to measure correlation between}

\item{primefeature}{one feature of interest to measure correlation against all other features}
}
\description{
Get Correlation Coefficient and its distribution of a single prime feature with all the other features
}
\examples{
transdf<- data_transform(df=GTEXv7[,3:ncol(GTEXv7)],transformation='log2', featurelist=GTEXv7[,2], medianthres=0.5)
inputdf<-transdf[[1]] #First item in the list returned is the transformed Data
primefcorr <- primefeature_corr(df=inputdf[,-ncol(inputdf)],featurelist=inputdf$feature ,primefeature="RNASEH2A");
}
