% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrs.R
\name{primefeature_corr}
\alias{primefeature_corr}
\title{Prime Feature Correlation
Get Correlation Coefficient and its distribution of a single prime feature with all the other features}
\usage{
primefeature_corr(df,featurelist,primefeature,corrmeth,quantiles)
}
\arguments{
\item{df}{A dataframe with echo row corresponding to a feature}

\item{featurelist}{listing of feature names or ids to measure correlation between}

\item{primefeature}{one feature of interest to measure correlation against all other features}

\item{corrmeth}{Correlation method used. 'pearson' or 'spearman' (default='pearson')}

\item{quantiles}{list of corresponding probabilities for quantiles to be obtained
(default = c(0,0.01,.01,.05,0.10,.25,.50,.75,.90,.95,.99,.999,1) )}
}
\description{
Prime Feature Correlation
Get Correlation Coefficient and its distribution of a single prime feature with all the other features
}
\examples{
transdf<- data_transform(df=GTEX[-1],transformation='log2',
                         featurelist=GTEX$Description, medianthres=1)
inputdf<-transdf[[1]] #First item in the list returned is the transformed Data
primefcorr <- primefeature_corr(df=inputdf,featurelist=rownames(inputdf),primefeature="RNASEH2A");

}
